Clean folder structure
Close: #29
